requirements.txt
Flask
Flask-Session
gunicorn
python-dotenv
redis
bcrypt
PyJWT
mysql-connector-python
pyyaml
ua-parser
user-agents
